paginate elastic response

+++ core-service/src/modules/search/usecase/search_ucase.go
@@ -20,7 +20,7 @@ func NewSearchUsecase(s domain.SearchRepository, timeout time.Duration) domain.S
 	}
 }
 
-func (n *searchUsecase) Fetch(c context.Context, params *domain.Request) (res []domain.SearchListResponse, tot int, err error) {
+func (n *searchUsecase) Fetch(c context.Context, params *domain.Request) (res []domain.SearchListResponse, tot int64, err error) {
 	ctx, cancel := context.WithTimeout(c, n.contextTimeout)
 	defer cancel()
